patch cv controllers and router
