--- src/server/api/controllers/cvs.controller.js
@@ -1,61 +1,0 @@
-/*  This is the controller for cvs */
-
-const knex = require('../../config/db');
-const Error = require('../lib/utils/http-error');
-const moment = require('moment-timezone');
-
-const getCvs = async () => {
-  try {
-    return await knex('cv').select('cv.id', 'cv.title');
-  } catch (error) {
-    return error.message;
-  }
-};
-
-const getCvById = async (id) => {
-  try {
-    const cvs = await knex('cv')
-      .select('cv.id as id', 'title')
-      .where({ id });
-    if (cvs.length === 0) {
-      throw new Error(`incorrect entry with the id of ${id}`, 404);
-    }
-    return cvs;
-  } catch (error) {
-    return error.message;
-  }
-};
-
-const editCv = async (cvId, updatedCv) => {
-  return knex('cv')
-    .where({ id: cvId })
-    .update({
-      title: updatedCv.title,
-      createdAt: moment(updatedCv.createdAt).format(),
-    });
-};
-
-const deleteCv = async (cvId) => {
-  return knex('cv')
-    .where({ id: cvId })
-    .del();
-};
-
-const createCv = async (body) => {
-  await knex('cv').insert({
-    title: body.title,
-    createdAt: moment(body.createdAt).format(),
-  });
-
-  return {
-    successful: true,
-  };
-};
-
-module.exports = {
-  getCvs,
-  getCvById,
-  deleteCv,
-  createCv,
-  editCv,
-};
