+++ src/server/api/routes/api-router.js
@@ -36,7 +36,7 @@ router.use('/documentation', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
 // Application routes
 router.use('/modules', modulesRouter);
 
-router.use('/cv', cvsRouter);
+router.use('/cvs', cvsRouter);
 
 router.use('/users', usersRouter);
